requirements.txt
apache-airflow
jupyter
pydantic
pytest
omegaconf
pre-commit
graphviz
pandas
pandera
pyarrow
mockito
dask
matplotlib

